Repair pagefreezer diff to return JSON data
It was returning the JSON as a string, even though the diff server is designed to handle JSON serializable objects. This was previous resulting in a JSON object with a value that was a JSON encoded string! Now we just have a nice simple JSON object.

Also stub out a test for this. I'm not sure right now how to best mock HTTP requests here, which is why this is only a stub. (@danielballan?)