Only raise request errors for actual IA errors
The IA servers respond with the original, stored request's status code, meaning a 404, 500, etc. status code actually be perfectly ok. To work around this, we look for the 'Memento-Datetime' header, which should probably be present on all valid responses. Also raise the underlying error from requests, which will include more useful information about the request and URL that had a problem than the assertion error did.