Remove 'Link: [url]' from visual diff output
This is done by creating our own version of `lxml.html.diff.htmldiff()` instead of calling it directly. The twist is that, after tokenizing the old and new HTML documents, we walk through the list of tokens and modify it as needed for better diffing or output. In this case, we just replace `href_token` instances with `MinimalHrefToken` instances, which, when rendered for final output, render nothing.