Handle undiffable content within the diff routine
This change removes our crazy hack for ripping out undiffable content before diffing and putting it back afterward and instead does it right: if an element is "undiffable" (meaning we can't diff *within* it), we treat it as a single, complete token to be diffed. If it has any changes, the whole thing will be changed as one. If it doesn't, it'll actually have a change to match up as unchanged (instead of *always* showing as a change like it did before).

This does make it intensely obvious how the flatten+tokenization steps need be brought together and handled more clearly. Having bits related to how a given element should be treated separated across those two different functions is bad.

Fixes #146