Add a rough way of diffing inside similar links
Instead of simply comparing two link listings against each other and leaving them as equal or not, allow two links to be "equal" if only their text or their href are the same. Then, when assembling the final diff, check whether "equal" items actually have the same hash, and, if not, use Diff-Match-Patch to diff their text and href.