Fix broken change_count in links diff
When I added the two-pass diffing here, I forgot that I was counting changes by evaluating the first pass, which misses any changes that are only found in the second pass. This is unfortunately a bit of a deoptimization, but I'm not sure what a more ideal solution would be. (And, practically speaking, it's not a particularly painful deoptimization).