Limit the number of spacer tokens we can add
This is SHORT TERM pain relief for #169. It's not a good solution, but it'll keep us hobbling along.

Waaaay back when I was still trying to tame lxml's differ from the outside without forking it, I added "spacer tokens" to the token stream that we diff for a document. The idea was to provide areas of sameness at the boundaries of certain kinds of elements, like list items, where we wanted to avoid (but not outright stop) insertions and deletions bleeding across from one element to the other. It worked alright and significantly improved some messy diffs.

Now we have some pages that take forever to diff because we add so many spacers that it completely bogs down SequenceMatcher. This is a *short-term* fix that simply caps the number of spacer tokens we can add at a constant amount (2500). Since we have fully forked the diff routine here from lxml, we have a lot of ways we can make smarter improvements than by using spacer tokens (and some of those have already been implemented). In the long run, we should look at better fundamental alternatives to the whole spacer token idea. For now, though, this simple cap helps keep the differ from suddenly spiking our memory and CPU, which can cause serious detriment to the whole platform.