Make CDXClient and MementoClient context managers
The clients needed `close()` methods to close the sessions they hold, so I went ahead and also made them context managers so doing so in actual code was a little nicer. I *didn't* reproduce the extra checks we had in my earlier, non-class based code, so this makes use the context manager with a user-managed session dangerous (the context manager might close it prematurely).