Use a single, long-lived process pool for diffs
The diffing server was constructing and using an independent process pool on every request. It seems unlikely that this is the main cause of our scalability issues (but maybe?), but it still seems like an obviously good thing to do. There's not much point in a pool if you aren't sharing it.

See also #303.