Upgrade from Raven to the new Sentry SDK
Raven is officially deprecated now (even though the new SDK is at v0.5, but ¯\_(ツ)_/¯), so the first thing to do in expanding our Sentry use is to upgrade! This also fixes the script to work with the current version of the `internetarchive` module. We must not have updated it since we changed that!