Reorganize remote content fetching in diff server (#431)
This resolves the deprecations in Tornado's HTTP client by reorganizing the code responsible for fetching and validating the content to diff. We previously had a lot of logic spread around `get()` that was just fetching the remote content. I've placed all of it into a method for fetching a single URL. Hopefully this also makes it a little more readable, though it could probably still be a bit cleaner.

This also adds tests for hash validation -- it turns out we didn't have any before!

Fixes #282.