Handle CDX API errors a little more gracefully (#430)
In the CDX search method, we previously just charged straight through situations where the HTTP response itself was an error and wound up raising an `UnexpectedResponseFormat` when attempting to parse the first line as CDX data. That means the user only ever sees the first line of the error, and is super confusing. Now we check for actual HTTP error codes right away and raise with a more complete message.