Use dateutil for date parsing, drop pandas (#509)
The only actual remaining use of pandas in our codebase was for parsing dates, which is a little overkill! We also already depended on dateutil, so we should actually have used it in all the places we are parsing dates. This does so, and since there are no remaining uses of pandas, drops it from the dependencies.

This also fixes our handling of non-UTC dates in CDX searches.

Fixes #360.