Add script to manually import annotations from analyst CSVs (#487)
This adds a script that will read a CSV matching the format of our analysts’ current sheets and create annotations in the database for each row. Run it like:

    scripts/annotations_import <PATH_TO_CSV>

Add the `--is_important_changes` option if the sheet represents “important” changes. This mainly affects how the `significance` field is calculated.

This is a component of edgi-govdata-archiving/web-monitoring#141