Use extension to guess MIME type for local files (#457)
I used the built in mimetypes package to get information required for the Content-Type and Content-Encoding headers. mimetypes only suggests Content-Encoding values for a few archive formats (eg .gz, .bz2), so that's not likely to be relevant for our purposes, but, hey, it's free.

`Content-Type` guessing logic and header setting now also lives entirely in the `MockResponse` class. If you pass custom headers to the constructor, it won’t overwrite your settings with a guess.