Handle broken process pool in diffing server well (#38)
We previously tried to reset the process pool that runs the actual diff routines every time it broke, and failed to handle the case where we exhausted all our retries (we'd wind up returning `None`, which is not an acceptable result). We now don't reset the pool and instead raise an exception on the last try, preventing weird errors from happening later on because we returned a bad value (`None`) from `DiffHandler.diff()`.

This also does a little more work to ensure that we aren't thrashing the process pool when a lot of diffs are happening simultaneously. We previously had each diff blindly reset the pool, which means that if multiple diffs were in flight when the pool broke, it would get reset multiple times, even though it really only needed to get reset once. This was probably wasting a lot of memory and CPU time.

Note that the tests here don't really test the first issue. There's not a good way to differentiate from the outside whether an unknown error was the process pool or something else.

Fixes #33.