HTML diff: Maintain spacing in text/tags
It turns out that some of the display breakage we've had with `html_render_diff` is happening because we are nicely reformatting the source on pages that use `white-space: pre;` or similar CSS styling to treat the spacing in the source code as significant. Solve this by working harder to maintain the original spacing and not unintentionally reformatting the source.

Fixes #36.