HTML diff: Maintain spacing in text/tags (#40)
It turns out that some of the display breakage we've had with `html_render_diff` is happening because we are nicely reformatting the source on pages that use `white-space: pre;` or similar CSS styling to treat the spacing in the source code as significant. For example, the following will display poorly, but it's what we were outputting:

    <!-- Before Version -->
    <div style="white-space: pre;">Some text</div>
    
    <!-- Before Version -->
    <div style="white-space: pre;"></div>
    
    <!-- Diff Output -->
    <div style="white-space: pre;">
      <del>
        Some text
      </del>
    </div>

Solve this by working harder to maintain the original spacing and not unintentionally reformatting the source, so we now output:

    <!-- Diff Output -->
    <div style="white-space: pre;"><del>Some text</del></div>

Fixes #36.