Compare images lazy-loaded with JS properly (#39)
When JS is used to lazy-load images, they do not have `src` or `srcset` attributes, which means we give the images an empty list of URLs. Unfortunately, we don't always know where the lazy-loaded image's URL is stored, so we can't solve this case perfectly, but the *vast* majority of the time, `data-src` and `data-srcset` are used, so that's what we fall back on here.

Additionally, if an image has no source URLs, that should still be OK, and we'll treat two images with no URLs as the same (we previously treated them differently).

Fixes #37.