Stop showing invisible content in HTML diff (#43)
This also solves an issue where text coloring would be incorrect and sometimes hard to read in Safari. This change is purely about the additional styling we add to diffs -- rather than use `all: unset` in our CSS, we take a more precise approach and only reset certain styles, and do not apply most of those resets to the descendents of the `<ins>`/`<del>` elements (we only apply them to the elements themselves).

Fixes #24.