Split docs requirements into separate group
On Python versions earlier than 3.10, flake8 and Sphinx are just not going to be compatible, and need to be in separate groups so both don't have to be installed at the same time or in the same environment. This also updates our ReadTheDocs setup to use their newer configuration file format. Overall, this follows the arrangement I set up for the Wayback package in https://github.com/edgi-govdata-archiving/wayback/pull/91.