Make sure all test HTTP requests are mocked (#159)
Some tests are currently failing in #158 because http://httpstat.us is down. We shouldn’t really be making any live requests in tests if we can avoid it, and this fixes that.

This adds a decorator that makes using `MockAsynchttpClient` a bit easier, adds error raising functionality to it, and makes sure it gets invoked everywhere it should be (a lot of missing callsites!).