Handle `binary/octet-stream` content type
Some recent additions to our upload script are now storing response bodies in S3 with the content type `binary/octet-stream`. This appears to be a change in boto3 (the AWS SDK), where it is now using that as a generic content type instead of `application/octet-stream`. This new type is not actually valid; I'm not sure why they're doing it, but this is functionally the same, and should not cause us to consider something as "definitely not HTML".

(Obviously we should fix the upload script, too, but that is a secondary concern.)