Limit number of diffs per worker process in server
This adds a new `MAX_DIFFS_PER_WORKER` environment variable that limits the number of diffs performed by a single worker. After a worker process performs this many diffs, it is shut down and replaced with a fresh process.

This is an ugly first cut and needs a lot of cleaning up. It also doesn't really account for things per worker -- it just restarts the pool after DIFFER_PARALLELISM * MAX_DIFFS_PER_WORKER diffs. But newer versions of Python have an API that does this right and which we should eventually switch to when compatible.

Fixes #202.