Fix innacurate XML prolog detection in server (#209)
The regular expression used for XML prolog detection (for character encoding declarations) was malformed. This wasn't a huge problem since it was just detecting too loosely and wouldn't have missed a prolog if present, but it does mean that it was probably finding false positives sometimes. This fixes the issue.