Make cchardet optional (#211)
Cchardet is not compatible with Python 3.11+ (there is an alpha release that works on 3.11 and 3.12, but it is now old too), and we need to make and keep this package easier to install, update, and secure. Cchardet seemed very dependable for a long time, but its maintainer eventually took a long break. It's pretty frozen, and we cannot depend on it as a hard requirement.

The detection test had to be amended to a different encoding in order to get something that is supported and reliably detected across all detector options.

Partially covers #196.